Polish HttpComponents WebClient implementation
